Remove support for partial keyword search
Currently, `find` command is not responsive (i.e., results do
not populate until user presses `Enter` key). In this scenario,
allowing a partial keyword search might lead to the user getting
confused.

Hence, this feature is being removed, and will be introduced at
a later iteration when it makes more sense to include it.