Remove check to see if file path refers to directory
Code to ensure that user provided a file path referring
to a file and not a directory was introduced in the recent
commit. This caused test cases to fail.

Upon examining the code, it was noted that Java automatically
handles this by treating it as part of the filename or
removing it completely (the behaviour is OS-dependent). Hence, the code
added was removed.